feat: support fractional TTW hours (e.g. 2.5 = 2h 30m)
- schema: ttwTargetHours Int? → Float?
- backend: parseInt → parseFloat for ttwTargetHours in POST/PUT
- form: step="0.5", min="0.5", placeholder updated to "e.g. 2.5"
- formatTTW: rewritten to display hours + minutes from fractional input

+++ backend/src/routes/articles.js
@@ -106,7 +106,7 @@ router.post(
           deadline:         deadline ? new Date(deadline) : null,
           briefNotes:       briefNotes || null,
           wordCountTarget:  wordCountTarget  ? parseInt(wordCountTarget)  : null,
-          ttwTargetHours:   ttwTargetHours   ? parseInt(ttwTargetHours)   : null,
+          ttwTargetHours:   ttwTargetHours   ? parseFloat(ttwTargetHours)   : null,
         },
         include: articleInclude,
       }),
